Made major changes to the repository
Included the microjson repository to the project to allow for processing
the csm.json file that is going to be used for the configurations of the
csm binary.

Created a library color-print-lib.c that basically makes it easier for
printing colored outputs on the command line and kept it as generic as
possible to make it usable basically anywhere.

Updated the csm-commandline-argument portion of the code to actually
process command line arguments and place the corresponing information to
a class that will hold the information simply.

Cleaned up some of the debug messages.

Added a template header file to make it easier for myself to create new
header files that have the correct guards for each one of the header
files that are there.

Updated the Makefile to actually generate the .d files that includes all
the dependencies for each one of the individual files. Also updated the
makefile to be able to process c and cpp files in the same way that the
cpp portion of the makefile does it. i.e. Only process and include the c
files that are specified and nothing else, so that we don't accidentally
compile and include a file that no longer belongs in the executable.