Huge changes to the repository
Added a new logging library to the application. At the moment the source
code in the repository in placed there by copying and pasting the code
manually from the logging repository into the csm source code to be
used. At some point this should be changed to be managed through the
makefile so that any updates to the logging library can be pulled in
automatically whenever the make file is run.

Added code in the makefile to make sure that files that just need to be
copied over for the project to work can just be copied over. This would
mean that the files that need to be copied over are copied over first
and then the application is compiled.

Updated the color-print-lib.c source to check the color_code that is
passed through to the function to make sure that an incorrect color code
isn't passed through. This is likely going to be checked again later to
make sure that it is working properly.

Moved the print dialogs to their own file.

Added a test json file for the application to work with.

Updated the template.h file to have the guards for c libraries in a c++
environment.

Added the first revision of the json-process-utils.c file that is meant
to help with the processing of large JSON files in low memory
environments. It's at a point where everything is compiling correctly
and there are no warnings and now it just needs to be tested to make
sure that the source code is running correctly and actually producing
the correct output.